Update README.MD; link to preivously fixed issues.
Updated the README.MD to reflect the fact that conversion of selected
text will now auto-extend to convert the entire first line and
last line selected.  This behavior was changed with the fix
for BrianGarland/vscode-rpgfree#90.

As my prior commits for the following issues did not correctly create
the link to the issues and (therfore did not) close them, I am
re-listing the issues fixed on prior commits:

* Fixes BrianGarland/vscode-rpgfree#32
* Fixes BrianGarland/vscode-rpgfree#35
* Fixes BrianGarland/vscode-rpgfree#45
* Fixes BrianGarland/vscode-rpgfree#48
* Fixes BrianGarland/vscode-rpgfree#50
* Fixes BrianGarland/vscode-rpgfree#61
* Fixes BrianGarland/vscode-rpgfree#62
* Fixes BrianGarland/vscode-rpgfree#81